Adding a license key:
This will be used from the client program to allow the possibility 
to add an host from the client it self entering that license.
We will be then able to link the host to a customer and to
a contract.

Should we limit the number of hosts valid for a contract ?